Big commit with plenty of features
- Following leader on shout
- Flocking algorithm (still need some adjustment)
- Attacking closest enemy (when not follow)
- Hp bar
- sprites changed to white to be better colorized programatically

For now I put that project on hold. After implementing most of
mvp features I realized that in current shape project is not fun.
For now I am still quite keen to test that idea but I believe that
it needs a lot of adjustment to make it fun.
RTS for beginning could waste a lot of my time and for now that
project is put on hold and I'm making step back for something
easier.